Move status utility to app context
