+++ shared/src/business/utilities/getFormattedTrialSessionDetails.test.js
@@ -1,6 +1,9 @@
 import { Case } from '../entities/cases/Case';
 import { MOCK_CASE } from '../../../../shared/src/test/mockCase';
-import { formattedTrialSessionDetails } from './getFormattedTrialSessionDetails';
+import {
+  formattedTrialSessionDetails,
+  getTrialSessionStatus,
+} from './getFormattedTrialSessionDetails';
 import { omit } from 'lodash';
 
 import { applicationContext } from '../../../../web-client/src/applicationContext';
@@ -300,5 +303,48 @@ describe('formattedTrialSessionDetails', () => {
       },
     });
     expect(result.computedStatus).toEqual('Closed');
+  });
+
+  describe('getTrialSessionStatus', () => {
+    it('returns `closed` when all trial session cases are inactive / removed from trial', () => {
+      const session = {
+        caseOrder: [
+          { docketNumber: '123-19', removedFromTrial: true },
+          { docketNumber: '234-19', removedFromTrial: true },
+        ],
+      };
+
+      const results = getTrialSessionStatus(session);
+
+      expect(results).toEqual('closed');
+    });
+
+    it('returns `open` when a trial session is calendared and does not meet conditions for `closed` status', () => {
+      const session = {
+        caseOrder: [
+          { docketNumber: '123-19' },
+          { docketNumber: '234-19', removedFromTrial: true },
+        ],
+        isCalendared: true,
+      };
+
+      const results = getTrialSessionStatus(session);
+
+      expect(results).toEqual('open');
+    });
+
+    it('returns `new` when a trial session is calendared and does not meet conditions for `closed` status', () => {
+      const session = {
+        caseOrder: [
+          { docketNumber: '123-19' },
+          { docketNumber: '234-19', removedFromTrial: true },
+        ],
+        isCalendared: false,
+      };
+
+      const results = getTrialSessionStatus(session);
+
+      expect(results).toEqual('new');
+    });
   });
 });
